requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.23
pydantic>=2.7
pydantic-settings>=2.2
httpx>=0.26
tenacity>=8.2
structlog>=24.1
python-multipart>=0.0.9
anyio>=4.0
beautifulsoup4>=4.12
lxml>=4.9
python-docx>=1.1
pypdf>=4.0
datasketch>=1.6
apscheduler>=3.10

# Optional (planned in later milestones)
SQLAlchemy>=2.0
alembic>=1.12
psycopg2-binary>=2.9 ; platform_system != "Windows"
qdrant-client>=1.8
orjson>=3.10

# Testing
pytest>=7.4
pytest-asyncio>=0.23

